requirements.txt
anyio==3.6.1
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
bidict==0.22.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SocketIO==5.2.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
python-telegram-bot==20.0a2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
selenium==4.4.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
Werkzeug==2.2.2
wsproto==1.1.0
